refactor(users): reuse client error parsing
